Cleaned up imports, and fixed all warnings. Now using doubles instead of floats for lossless converions to int

+++ IEntity.h
@@ -1,13 +1,12 @@
 #pragma once
 
-
-// forward declaration
 class GameWorld;
+
 
 // class definition
 class IEntity {
 
 public:
-	virtual void Init(GameWorld* world, float startX, float startY, float velX, float velY) = 0;
+	virtual void Init(GameWorld* world, double startX, double startY, double velX, double velY) = 0;
 	virtual void Update(float deltaTime) = 0;
 };
